pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-text-search"
version = "0.1.0"
description = "A local semantic search CLI for markdown and text note vaults."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
  "qdrant-client>=1.11,<2.0",
  "rich>=13.7,<14.0",
  "tomli-w>=1.0,<2.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
]

[project.scripts]
local-text-search = "local_text_search.cli:run"

[tool.setuptools.packages.find]
include = ["local_text_search*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
